+++ showcase/03-songbird-discovery/03-auto-advertise/demo.sh
@@ -91,7 +91,7 @@ REGISTER_PAYLOAD='{
     "proof-generation"
   ],
   "metadata": {
-    "version": "0.8.9",
+    "version": "0.9.14",
     "startup_time": "'$(date -Iseconds)'",
     "auto_registered": true
   }